feat: swap InMemoryNode::seek_until to use a VecDeque
This is replacing the existing implemenation, which uses a double ended iterator.

This is being done so that CursorSeek entries that don't consume at
least one char can "re-push" that char back on to the deque and
reprocess it a second time after (for example) an `until_fn` is pushed.