feat: collapse into one remove_nodes_sequentially_until call in Selection::delete to get rid of issue
The issue here seems to be that the first remove_nodes_sequentially_until call breaks the node
hierarchy, so that when it is called a second time, the nodes that are
trying to be traversed no longer are in the hierarchy, which makes the depth
calculations wrong (end up as 0), which leds to incorrect results. So,
move to a single call to fix this.