fix: address issue where InMemoryNode::remove_nodes_sequentially_until would remove a node's children when just the node was to be deleted
Now, what happens is if a node has children, it's replaced with this new
node called "Fragment" (which has no literal value). If at the end of
going through all a Fragment's children it is empty, THEN it gets
deleted. This means that if the code loops halfway through the children,
then the fragment will persist, keeping the hierarchy around.